refactor: decompose llvm_codegen into functional modules
Split the monolithic `llvm_codegen.rs` into specialized submodules for better organization and maintainability.

Changes:
- **Module Structure**:
  - `llvm_temporary/llvm_codegen/` (New directory)
    - `mod.rs`: Re-exports public functions.
    - `ir.rs`: Core AST traversal and module/function setup (`generate_ir`).
    - `address.rs`: Memory address calculation (`generate_address_ir`).
    - `types.rs`: Type conversion (`wave_type_to_llvm_type`) and `VariableInfo`.
    - `consts.rs`: Constant creation (`create_llvm_const_value`).
    - `format.rs`: Format string parsing (`wave_format_to_c`, `wave_format_to_scanf`).
    - `legacy.rs`: Deprecated `TokenType`-based functions (kept for compatibility).
- **Refactoring**:
  - Moved relevant functions from `llvm_codegen.rs` to their respective new files.
  - Updated `wave_format_to_c` signature to accept `Context` reference (required for float type checking).
  - Updated `gen_print_format_ir` in `statement/io.rs` to pass `context` to `wave_format_to_c`.
- **Cleanup**: Removed the original single `llvm_codegen.rs` file.

This modularization clarifies the responsibilities of the codegen backend and simplifies future extensions.
